Refactor gitbook_worker into modules
