+++ tools/gitbook_worker/tests/test_additional.py
@@ -1,4 +1,4 @@
-import gitbook_worker
+from gitbook_worker import utils, linkcheck, source_extract, validate_metadata
 
 class DummyResponse:
     def __init__(self, status_code=200, reason="OK"):
@@ -15,13 +15,13 @@ def test_parse_summary(tmp_path):
     chapter1.write_text("c1")
     chapter2.write_text("c2")
     summary.write_text("* [C1](chapter1.md)\n  * [C2](sub/chapter2.md)\n")
-    result = gitbook_worker.parse_summary(str(summary))
+    result = utils.parse_summary(str(summary))
     assert result == [str(chapter1), str(chapter2)]
 
 
 def test_extract_multiline_list_items():
     text = "1. first\ncontinued\n2- second\n* bullet\n  more bullet"
-    items = gitbook_worker.extract_multiline_list_items(text)
+    items = source_extract.extract_multiline_list_items(text)
     assert items == [
         "1. first\ncontinued",
         "2- second",
@@ -34,7 +34,7 @@ def test_validate_metadata(tmp_path):
     good.write_text("---\ntitle: T\nauthor: A\ndate: 2021\n---\n")
     bad = tmp_path / "bad.md"
     bad.write_text("---\ntitle: T\ndate: 2021\n---\n")
-    issues = gitbook_worker.validate_metadata([str(good), str(bad)])
+    issues = validate_metadata([str(good), str(bad)])
     assert (str(bad), "Missing metadata field: author") in issues
     assert not any(i[0] == str(good) for i in issues)
 
@@ -44,14 +44,14 @@ def test_check_duplicate_headings(tmp_path):
     f2 = tmp_path / "b.md"
     f1.write_text("# Title\n")
     f2.write_text("# Title\n")
-    dup = gitbook_worker.check_duplicate_headings([str(f1), str(f2)])
+    dup = linkcheck.check_duplicate_headings([str(f1), str(f2)])
     assert dup == [(str(f2), 1, "title", f"{f1}:1")]
 
 
 def test_list_todos(tmp_path):
     md = tmp_path / "todo.md"
     md.write_text("line\n# TODO: task\ncontent\n# FIXME something")
-    todos = gitbook_worker.list_todos([str(md)])
+    todos = linkcheck.list_todos([str(md)])
     assert len(todos) == 2
     assert todos[0][1] == 2
 
@@ -67,8 +67,8 @@ def test_check_images(tmp_path, monkeypatch):
     )
     def fake_head(url, timeout=5):
         return DummyResponse(404, "Not Found")
-    monkeypatch.setattr("gitbook_worker.requests.head", fake_head)
-    missing = gitbook_worker.check_images([str(md)])
+    monkeypatch.setattr(linkcheck.requests, "head", fake_head)
+    missing = linkcheck.check_images([str(md)])
     paths = [m[2] for m in missing]
     assert str(tmp_path / "missing.png") in paths
     assert "https://example.com/x.png" in paths
