+++ tools/gitbook_worker/tests/test_external_repair.py
@@ -1,4 +1,4 @@
-import gitbook_worker
+from gitbook_worker import ai_tools, source_extract
 
 
 def test_proof_and_repair_replaces_content(tmp_path, monkeypatch):
@@ -12,10 +12,10 @@ def test_proof_and_repair_replaces_content(tmp_path, monkeypatch):
     def fake_proof(**kwargs):
         return True, {"success": True, "new": "1. Example NEW", "validation_date": "2024-01-01", "type": "external reference", "hint": None}
 
-    monkeypatch.setattr(gitbook_worker, "extract_sources_of_a_md_file_to_dict", fake_extract)
-    monkeypatch.setattr(gitbook_worker, "proof_and_repair_external_reference", fake_proof)
+    monkeypatch.setattr(source_extract, "extract_sources_of_a_md_file_to_dict", fake_extract)
+    monkeypatch.setattr(ai_tools, "proof_and_repair_external_reference", fake_proof)
 
-    gitbook_worker.proof_and_repair_external_references([str(md)], "", "", "", "")
+    ai_tools.proof_and_repair_external_references([str(md)], "", "", "", "")
 
     content = md.read_text().splitlines()
     assert content.count("1. Example NEW") == 1
