+++ tools/gitbook_worker/tests/test_more_functions.py
@@ -1,5 +1,5 @@
 import os
-import gitbook_worker
+from gitbook_worker import utils, linkcheck, source_extract, validate_metadata
 
 class DummyResponse:
     def __init__(self, status=200, reason="OK"):
@@ -19,27 +19,27 @@ def test_parse_summary_nested(tmp_path):
 * [One](chap1.md)
   * [Two](nested/level/chap2.md)
 """)
-    result = gitbook_worker.parse_summary(str(summary))
+    result = utils.parse_summary(str(summary))
     assert result == [str(f1), str(f2)]
 
 
 def test_extract_multiline_list_items_alpha():
     text = "a) Apple\nb) Banana\n  extra\n1. Numbered"
-    items = gitbook_worker.extract_multiline_list_items(text)
+    items = source_extract.extract_multiline_list_items(text)
     assert items == ["a) Apple", "b) Banana\n  extra", "1. Numbered"]
 
 
 def test_validate_metadata_no_frontmatter(tmp_path):
     md = tmp_path / "file.md"
     md.write_text("content only")
-    issues = gitbook_worker.validate_metadata([str(md)])
+    issues = validate_metadata([str(md)])
     assert issues == []
 
 
 def test_check_duplicate_headings_same_file(tmp_path):
     md = tmp_path / "dup.md"
     md.write_text("# Head\ntext\n## Other\n# Head\n")
-    dups = gitbook_worker.check_duplicate_headings([str(md)])
+    dups = linkcheck.check_duplicate_headings([str(md)])
     assert dups == [(str(md), 4, "head", f"{md}:1")]
 
 
@@ -48,7 +48,7 @@ def test_list_todos_multi(tmp_path):
     b = tmp_path / "b.md"
     a.write_text("TODO first\n")
     b.write_text("text\nFIXME second\n")
-    todos = gitbook_worker.list_todos([str(a), str(b)])
+    todos = linkcheck.list_todos([str(a), str(b)])
     assert [(os.path.basename(t[0]), t[1]) for t in todos] == [("a.md", 1), ("b.md", 2)]
 
 
@@ -63,8 +63,8 @@ def test_check_images_errors(tmp_path, monkeypatch):
     def fake_head(url, timeout=5):
         raise Exception("boom")
 
-    monkeypatch.setattr(gitbook_worker.requests, "head", fake_head)
-    missing = gitbook_worker.check_images([str(md)])
+    monkeypatch.setattr(linkcheck.requests, "head", fake_head)
+    missing = linkcheck.check_images([str(md)])
     paths = [m[2] for m in missing]
     assert os.path.join(str(tmp_path), "missing.png") in paths
     assert "http://remote/img" in paths
