+++ tools/gitbook_worker/tests/test_quality_checks.py
@@ -1,5 +1,13 @@
 import os
-import gitbook_worker
+from gitbook_worker import (
+    utils,
+    linkcheck,
+    source_extract,
+    validate_metadata,
+    readability_report,
+    proof_and_repair_internal_references,
+    ai_tools,
+)
 
 class DummyResponse:
     def __init__(self, status_code=200, reason="OK"):
@@ -16,20 +24,20 @@ def test_parse_summary(tmp_path):
     summary = tmp_path / "SUMMARY.md"
     summary.write_text("* [One](chap1.md)\n* [Two](sub/chap2.md)\n")
 
-    result = gitbook_worker.parse_summary(str(summary))
+    result = utils.parse_summary(str(summary))
     assert result == [str(md1), str(md2)]
 
 
 def test_extract_multiline_list_items():
     text = "1. First line\ncontinued\n* Bullet\nmore bullet\n"
-    items = gitbook_worker.extract_multiline_list_items(text)
+    items = source_extract.extract_multiline_list_items(text)
     assert items == ["1. First line\ncontinued", "* Bullet\nmore bullet\n"]
 
 
 def test_validate_metadata(tmp_path):
     md = tmp_path / "meta.md"
     md.write_text("---\ntitle: T\nauthor: A\n---\nbody")
-    issues = gitbook_worker.validate_metadata([str(md)])
+    issues = validate_metadata([str(md)])
     assert issues == [(str(md), "Missing metadata field: date")]
 
 
@@ -38,14 +46,14 @@ def test_check_duplicate_headings(tmp_path):
     md2 = tmp_path / "b.md"
     md1.write_text("# Title\n")
     md2.write_text("## Title\n")
-    dups = gitbook_worker.check_duplicate_headings([str(md1), str(md2)])
+    dups = linkcheck.check_duplicate_headings([str(md1), str(md2)])
     assert dups and dups[0][0] == str(md2)
 
 
 def test_list_todos(tmp_path):
     md = tmp_path / "todo.md"
     md.write_text("Line\n# TODO fix\nText\n# FIXME later\n")
-    todos = gitbook_worker.list_todos([str(md)])
+    todos = linkcheck.list_todos([str(md)])
     assert [t[1] for t in todos] == [2, 4]
 
 
@@ -65,9 +73,9 @@ def test_check_images(tmp_path, monkeypatch):
     def fake_head(url, timeout=5):
         return responses[url]
 
-    monkeypatch.setattr(gitbook_worker.requests, "head", fake_head)
+    monkeypatch.setattr(linkcheck.requests, "head", fake_head)
 
-    result = gitbook_worker.check_images([str(md)])
+    result = linkcheck.check_images([str(md)])
 
     missing_paths = [r[2] for r in result]
     assert os.path.join(str(tmp_path), "missing.png") in missing_paths
@@ -80,21 +88,21 @@ def test_extract_sources_and_readability(tmp_path):
     md1.write_text("# T\n\n## Sources\n1. Ref https://ex.com\n")
     md2 = tmp_path / "two.md"
     md2.write_text("No sources here")
-    sources = gitbook_worker.extract_sources_to_dict([str(md1), str(md2)])
+    sources = source_extract.extract_sources_to_dict([str(md1), str(md2)])
     assert str(md1) in sources and sources[str(md1)]
 
     csv = tmp_path / "out.csv"
-    gitbook_worker.extract_sources([str(md1)], str(csv))
+    source_extract.extract_sources([str(md1)], str(csv))
     assert csv.exists()
 
-    report = gitbook_worker.readability_report([str(md2)])
+    report = readability_report([str(md2)])
     assert report[0][0] == str(md2)
 
 
 def test_citation_numbering(tmp_path):
     md = tmp_path / "cit.md"
     md.write_text("1. A\n3. B\n")
-    gaps = gitbook_worker.check_citation_numbering([str(md)])
+    gaps = linkcheck.check_citation_numbering([str(md)])
     assert gaps == [(str(md), [2])]
 
 def test_internal_reference_noop(tmp_path, monkeypatch):
@@ -102,8 +110,9 @@ def test_internal_reference_noop(tmp_path, monkeypatch):
     summary.write_text("* [A](a.md)\n")
     md = tmp_path / "a.md"
     md.write_text("text\n")
-    monkeypatch.setattr(gitbook_worker, "extract_sources_of_a_md_file_to_dict", lambda m: {})
-    gitbook_worker.proof_and_repair_internal_references([str(md)], str(summary))
+    monkeypatch.setattr(source_extract, "extract_sources_of_a_md_file_to_dict", lambda m: {})
+    monkeypatch.setattr(ai_tools, "extract_sources_of_a_md_file_to_dict", lambda m: {})
+    proof_and_repair_internal_references([str(md)], str(summary))
     assert md.read_text() == "text\n"
 
 def test_internal_reference_adds_footnote(tmp_path, monkeypatch):
@@ -115,6 +124,7 @@ def test_internal_reference_adds_footnote(tmp_path, monkeypatch):
     def fake_extract(file):
         return {str(file): [{"Ref": {"lineno": 1, "line": "1. Ref missing", "level": 1, "link": "missing.md", "numbering": "1"}}]}
 
-    monkeypatch.setattr(gitbook_worker, "extract_sources_of_a_md_file_to_dict", fake_extract)
-    gitbook_worker.proof_and_repair_internal_references([str(md)], str(summary))
+    monkeypatch.setattr(source_extract, "extract_sources_of_a_md_file_to_dict", fake_extract)
+    monkeypatch.setattr(ai_tools, "extract_sources_of_a_md_file_to_dict", fake_extract)
+    proof_and_repair_internal_references([str(md)], str(summary))
     assert "Ref" in md.read_text()
